added undo logic
